Split off statistics functionality to dashboard_statistic.py, fixed docstring error.
