+++ pydash/pydash_app/dashboard/aggregator/statistics.py
@@ -117,6 +117,7 @@ class TotalVisits(Statistic):
 
 
 class TotalExecutionTime(FloatStatisticABC):
+    """Total execution time in ms. Rendered value is rounded to 3 decimal places by default."""
     def should_be_rendered(self):
         return True
 
@@ -136,7 +137,7 @@ class TotalExecutionTime(FloatStatisticABC):
 
 
 class AverageExecutionTime(FloatStatisticABC):
-    """Keeps track of the average execution time of all endpoints that have been appended to it.
+    """Keeps track of the average execution time in ms of all endpoints that have been appended to it.
     Rendered value is rounded to 3 decimal places by default."""
     dependencies = [TotalVisits, TotalExecutionTime]
