+++ pydash/pydash_web/controller/__init__.py
@@ -13,6 +13,7 @@ from .change_password import change_password
 
 from .dashboards import dashboards
 from .dashboard import dashboard
+from .dashboard_statistic import dashboard_statistic
 from .register_dashboard import register_dashboard
 from .change_dashboard_settings import change_dashboard_settings
 from .delete_dashboard import delete_dashboard
